Changed menu/edited tariffs.
