+++ src/main/java/web/bean/tariff/TariffOverviewBean.java
@@ -3,6 +3,7 @@ package web.bean.tariff;
 import main.domain.Tariff;
 import main.service.TariffService;
 import web.core.helper.FrontendHelper;
+import web.core.helper.RedirectHelper;
 
 import javax.annotation.PostConstruct;
 import javax.faces.view.ViewScoped;
@@ -21,6 +22,9 @@ public class TariffOverviewBean implements Serializable {
     private List<Tariff> tariffs;
     private List<Tariff> filteredTariffs;
 
+    private double tariffInEuro;
+    private boolean ridingDuringRushHour;
+
     @PostConstruct
     public void init() {
         this.tariffs = this.tariffService.findAll();
@@ -38,6 +42,27 @@ public class TariffOverviewBean implements Serializable {
         FrontendHelper.displaySuccessSmallMessage("Succes!", "Tarief is succesvol gewijzigd.");
     }
 
+    public void create() {
+        if (tariffInEuro != 0) {
+            Tariff newTariff = new Tariff(this.tariffInEuro, this.ridingDuringRushHour);
+            tariffService.createOrUpdate(newTariff);
+            this.tariffs = tariffService.findAll();
+            FrontendHelper.displaySuccessSmallMessage("Het tarief is succesvol toegevoegd!");
+        } else {
+            FrontendHelper.displayErrorSmallMessage("Vul alstublieft een tarief in.");
+        }
+    }
+
+    public void remove(Tariff tariff) {
+        if (tariff != null) {
+            tariffService.delete(tariff);
+            this.tariffs = tariffService.findAll();
+            FrontendHelper.displaySuccessSmallMessage("Het tarief is succesvol verwijderd!");
+        } else {
+            FrontendHelper.displayErrorSmallMessage("Er ging iets mis.", "Probeer het opnieuw.");
+        }
+    }
+
     //<editor-fold defaultstate="collapsed" desc="Getters/Setters">
     public List<Tariff> getTariffs() {
         return tariffs;
@@ -53,6 +78,22 @@ public class TariffOverviewBean implements Serializable {
 
     public void setFilteredTariffs(List<Tariff> filteredTariffs) {
         this.filteredTariffs = filteredTariffs;
+    }
+
+    public double getTariffInEuro() {
+        return tariffInEuro;
+    }
+
+    public void setTariffInEuro(double tariffInEuro) {
+        this.tariffInEuro = tariffInEuro;
+    }
+
+    public boolean isRidingDuringRushHour() {
+        return this.ridingDuringRushHour;
+    }
+
+    public void setRidingDuringRushHour(boolean ridingDuringRushHour) {
+        this.ridingDuringRushHour = ridingDuringRushHour;
     }
     //</editor-fold>
 }
