--- src/main/webapp/pages/tariff/add.xhtml
@@ -1,47 +1,0 @@
-<ui:composition xmlns="http://www.w3.org/1999/xhtml"
-                xmlns:ui="http://java.sun.com/jsf/facelets"
-                xmlns:p="http://primefaces.org/ui"
-                xmlns:h="http://xmlns.jcp.org/jsf/html"
-                template="/templates/template.xhtml">
-
-    <ui:define name="title">Nieuw Tarief</ui:define>
-
-    <ui:define name="content">
-
-        <div class="container">
-
-            <header class="header">
-                <h1>Nieuw tarief</h1>
-                <h5 class="text-muted">Voeg een nieuw tarief toe</h5>
-            </header>
-
-            <div class="content-container">
-
-                <h:form>
-                    <div class="form-group">
-                        <p:inputText id="tariff"
-                                     placeholder="0,00"
-                                     required="true"
-                                     requiredMessage="Vul het tarief in Euro's in."
-                                     styleClass="form-control"
-                                     value="#{tariffBean.tariffInEuro}"/>
-                    </div>
-
-                    <div>
-                        <h:selectBooleanCheckbox value="#{tariffBean.ridingDuringRushHour}"/>
-                        <h:outputText value="Geldt dit tarief tijdens spits uren?"/>
-                    </div>
-                    <div>
-                        <p:commandButton action="#{tariffBean.create}"
-                                         id="create"
-                                         value="Nieuw"
-                                         styleClass="btn btn-success"/>
-                    </div>
-                </h:form>
-
-            </div>
-
-        </div>
-
-    </ui:define>
-</ui:composition>
