+++ src/main/webapp/pages/tariff/overview.xhtml
@@ -18,13 +18,13 @@
 
             <div class="content-container">
 
-                <h:form id="form">
+                <h:form id="tariffOverviewForm">
 
                     <p:dataTable editable="true"
                                  editMode="cell"
                                  emptyMessage="Er zijn geen tarieven gevonden die voldoen aan de criteria"
                                  filteredValue="#{tariffOverviewBean.filteredTariffs}"
-                                 id="tariffTable"
+                                 id="tariffDataTable"
                                  paginator="true"
                                  paginatorPosition="bottom"
                                  paginatorTemplate="{CurrentPageReport} {FirstPageLink} {PreviousPageLink} {PageLinks} {NextPageLink} {LastPageLink} {RowsPerPageDropdown}"
@@ -58,9 +58,9 @@
                         </p:column>
 
                         <p:column styleClass="text-align-center" width="10%">
-                            <p:commandButton action="#{tariffBean.remove(tariff)}"
+                            <p:commandButton action="#{tariffOverviewBean.remove(tariff)}"
                                              id="remove"
-                                             update="tariffTable"
+                                             update="tariffDataTable"
                                              value="Delete"
                                              styleClass="btn btn-danger">
                                 <p:confirm header="Weet u het zeker?" message="Wil u dit tarief echt verwijderen?"/>
@@ -73,6 +73,43 @@
             </div>
 
         </div>
+        <div class="container">
+
+            <header class="header">
+                <h1>Nieuw tarief</h1>
+                <h5 class="text-muted">Voeg een nieuw tarief toe</h5>
+            </header>
+
+            <div class="content-container">
+
+                <h:form id="newTariffForm">
+                    <div class="form-group">
+                        <p:inputText id="tariff"
+                                     placeholder="0,00"
+                                     required="true"
+                                     requiredMessage="Vul het tarief in Euro's in."
+                                     styleClass="form-control"
+                                     value="#{tariffOverviewBean.tariffInEuro}"/>
+                    </div>
+
+                    <div>
+                        <h:selectBooleanCheckbox value="#{tariffOverviewBean.ridingDuringRushHour}"/>
+                        <h:outputText value="Geldt dit tarief tijdens spits uren?"/>
+                    </div>
+                    <div>
+                        <p:commandButton action="#{tariffOverviewBean.create}"
+                                         id="create"
+                                         value="Nieuw"
+                                         process="newTariffForm"
+                                         update="tariffOverviewForm:tariffDataTable"
+                                         styleClass="btn btn-success"/>
+                    </div>
+                </h:form>
+
+            </div>
+
+        </div>
+
 
     </ui:define>
 </ui:composition>
